Simplifie structure and apply @skonefal code review
