+++ pkg/experiment/metadata.go
@@ -1,0 +1,8 @@
+package experiment
+
+// Metadata contains values which are identifying SwanMetrics.
+type Metadata struct {
+	ExperimentID string
+	PhaseID      string
+	RepetitionID int
+}
