+++ pkg/experiment/sensitivity/uploader.go
@@ -1,6 +1,6 @@
-package metrics
+package sensitivity
 
 // Uploader is a interface for SwanMetrics uploading into external media(like database or telemetry framework)
 type Uploader interface {
-	SendMetrics(Swan) error
+	SendMetadata(Metadata) error
 }
